# Utilize discriminant_value for more efficient deriving