@pcwalton 2015-06-13
Do you know if this improves compile times of rustc?