@Marwes 2015-06-13
@pcwalton I haven't tested the compile times so I can't say for certain. C-like enums generate a lot better IR which should help make both compiled programs faster as well as compiling that code faster. Other enums should get a less noticeable speedup as well.

I might do some benchmarking on the compiler tomorrow to see if it makes a difference though.