@bluss 2015-06-14
Does this affect codegen for `Option<T>`? Just thinking of our most important enum.