@Marwes 2015-06-14
I ran the same benchmark on an Option<i32> and I got a 1.1x speedup. So I'd guess it helps a little though it might very well vary depending on what type the Option contains.