@Marwes 2015-06-14
@pcwalton I compiled rustc a few times with and without this PR but I got to much noise to be able to say anything.