@Marwes 2015-06-14
Removed the comment mentioning #15523 as per @bluss suggestion, as that had been fixed.