@pcwalton 2015-06-16
@bors: r+